Switch to modulo operator instead of a mask
This way buffer can be of an arbitrary size. The performance does not
drop significantly and furthermore code is more readable.

The table below shows the number of instructions of different functions
on x86 and ARM architecture. Please note that in most cases performance
of queue and dequeue functions is the most important, as these are
going to be called more often than the other:

|       function       | modulo (x86) | no-modulo (x86) | modulo (ARM) | no-modulo (ARM) |
|----------------------|--------------|-----------------|--------------|-----------------|
| ring_buffer_queue    |      14      |       18        |      15      |       17        |
| ring_buffer_dequeue  |      15      |       15        |      17      |       17        |
| ring_buffer_peek     |      19      |       17        |      19      |       16        |
| all functions        |      128     |       111       |      132     |       114       |

Link below enables detailed analysis:
https://godbolt.org/z/crT893G13

Based on the above results, we may conclude that in most of the use
cases, modulo is at least as good as masking if not even better.